Fix flakiness in httpServer graceful shutdown integration test (#121749)
* Fix flakiness in httpServer graceful shutdown integration test

* lint
